requirements.txt
Flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
chardet==5.2.0
pycryptodome==3.17
xhtml2pdf==0.2.8
requests==2.31.0
python-dotenv==1.0.1
groq==0.1.0  # Use the correct version of the Groq Python client